+++ views/login.html
@@ -1,28 +1,28 @@
 <!doctype html>
-<html lang="en">
+<html lang="es">
 
 <head>
-    <title>Title</title>
+    <title>Inicio sesi&oacuten</title>
     <!-- Required meta tags -->
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
 
     <!-- Bootstrap CSS -->
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css"
-        integrity="sha384-9aIt2nRpC12Uk9gS9baDl411NQApFmC26EwAOH8WgZl5MYYxFfc+NcPb1dKGj7Sk" crossorigin="anonymous">
+    integrity="sha384-9aIt2nRpC12Uk9gS9baDl411NQApFmC26EwAOH8WgZl5MYYxFfc+NcPb1dKGj7Sk" crossorigin="anonymous">
     <link rel="stylesheet" href="../css/login.css">
 
 </head>
 
 <body class="text-center">
     <form class="form-signin">
-        <img class="mb-4 logo" src="../images/login/leaf.png" alt="" width="72" height="72">
-        <h1 class="h3 mb-3 font-weight-normal">Bienvenido!</h1>
-        <h1 class="h5 mb-3 font-weight-normal">Porfavor Iniciar Sesion</h2>
+        <img class="mb-4 logo" src="../images/leaf.png" alt="" width="72" height="72">
+        <h1 class="h3 mb-3 font-weight-normal">¡Bienvenido!</h1>
+        <h1 class="h5 mb-3 font-weight-normal">Por favor, inicia sesi&oacuten</h2>
         <input type="email" id="inputEmail" class="form-control" placeholder="Correo Electronico" required autofocus>
         <input type="password" id="inputPassword" class="form-control" placeholder="Contraseña" required>
-        <a href="">Olvide Mi Contraseña</a>
-        <button class="btn btn-lg btn-primary btn-block" type="submit">Iniciar Sesion</button>
+        <a href="">Olvid&eacute mi contraseña</a>
+        <button class="btn btn-lg btn-primary btn-block" type="submit">Iniciar Sesi&oacuten</button>
         <p class="mt-5 mb-3 text-muted">&copy; 2020</p>
     </form>
     <!-- Optional JavaScript -->
