+++ views/mi-cuenta.html
@@ -1,8 +1,8 @@
 <!doctype html>
-<html lang="en">
+<html lang="es">
 
 <head>
-    <title>Title</title>
+    <title>Mi Cuenta</title>
     <!-- Required meta tags -->
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
@@ -20,7 +20,7 @@
 <body>
     <header>
         <nav class="navbar navbar-expand-lg fixed-top navbar-dark">
-            <a class="navbar-brand" href="#">TuCV</a>
+            <a class="navbar-brand" href="#">Tu CV</a>
             <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
                 <span class="navbar-toggler-icon"></span>
             </button>
@@ -43,18 +43,19 @@
                         <a class="nav-link" href="experiencia.html">Experiencia</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="formacion.html">Formacion</a>
+                        <a class="nav-link" href="formacion.html">Formaci&oacuten</a>
                     </li>
                 </ul>
                 <ul class="navbar-nav ml-auto">
+                    <!-- Esto va solo pal admin
                     <li class="nav-item">
-                        <a class="nav-link" href="usuarios.html"><i class="fas fa-users"></i> Usuarios</a>
-                    </li>
+                    <a class="nav-link" href="usuarios.html"><i class="fas fa-users"></i> Usuarios</a>
+                    </li> -->
                     <li class="nav-item active">
                         <a class="nav-link" href="mi-cuenta.html"><i class="fas fa-user-circle"></i> Mi Cuenta</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="login.html"><i class="fas fa-sign-out-alt"></i> Cerrar Sesion</a>
+                        <a class="nav-link" href="login.html"><i class="fas fa-sign-out-alt"></i> Cerrar Sesi&oacuten</a>
                     </li>
                 </ul>
             </div>
@@ -65,7 +66,7 @@
         <form>
             <div class="form-row">
                 <div class="form-group col-md-6">
-                    <label for="inputCorreo">Correo Electronico:</label>
+                    <label for="inputCorreo">Correo Electr&oacutenico:</label>
                     <input type="text" class="form-control" id="inputCorreo" value="ABC@XYC.com" readonly>
                 </div>
             </div>
