+++ views/mis-datos.html
@@ -1,8 +1,8 @@
 <!doctype html>
-<html lang="en">
+<html lang="es">
 
 <head>
-    <title>Title</title>
+    <title>Mis Datos</title>
     <!-- Required meta tags -->
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
@@ -20,7 +20,7 @@
 <body>
     <header>
         <nav class="navbar navbar-expand-lg fixed-top navbar-dark">
-            <a class="navbar-brand" href="#">TuCV</a>
+            <a class="navbar-brand" href="#">Tu CV</a>
             <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
                 <span class="navbar-toggler-icon"></span>
             </button>
@@ -43,7 +43,7 @@
                         <a class="nav-link" href="experiencia.html">Experiencia</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="formacion.html">Formacion</a>
+                        <a class="nav-link" href="formacion.html">Formaci&oacuten</a>
                     </li>
                 </ul>
                 <ul class="navbar-nav ml-auto">
@@ -51,7 +51,7 @@
                         <a class="nav-link" href="mi-cuenta.html"><i class="fas fa-user-circle"></i> Mi Cuenta</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="login.html"><i class="fas fa-sign-out-alt"></i> Cerrar Sesion</a>
+                        <a class="nav-link" href="login.html"><i class="fas fa-sign-out-alt"></i> Cerrar Sesi&oacuten</a>
                     </li>
                 </ul>
             </div>
@@ -72,7 +72,7 @@
             </div>
             <div class="form-row">
                 <div class="form-group col-md-6">
-                    <label for="inputTelefono">Telefono Celular</label>
+                    <label for="inputTelefono">Tel&eacutefono Celular</label>
                     <input type="text" class="form-control" id="inputTelefono">
                 </div>
                 <div class="form-group col-md-6">
@@ -94,6 +94,7 @@
     <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
     <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js" integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo" crossorigin="anonymous"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.min.js" integrity="sha384-OgVRvuATP1z7JjHLkuOU7Xw704+h835Lr+6QL9UvYjZE3Ipu6Tp75j7Bh/kR0JKI" crossorigin="anonymous"></script>
+    <script src="../js/mis-datos.js"></script>
 </body>
 
 </html>
