+++ views/usuarios.html
@@ -1,8 +1,8 @@
 <!doctype html>
-<html lang="en">
+<html lang="es">
 
 <head>
-    <title>Title</title>
+    <title>Usuarios</title>
     <!-- Required meta tags -->
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
@@ -21,7 +21,7 @@
 <body>
     <header>
         <nav class="navbar navbar-expand-lg fixed-top navbar-dark">
-            <a class="navbar-brand" href="#">TuCV</a>
+            <a class="navbar-brand" href="#">Tu CV</a>
             <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
                 <span class="navbar-toggler-icon"></span>
             </button>
@@ -44,7 +44,7 @@
                         <a class="nav-link" href="experiencia.html">Experiencia</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="formacion.html">Formacion</a>
+                        <a class="nav-link" href="formacion.html">Formaci&oacuten</a>
                     </li>
                 </ul>
                 <ul class="navbar-nav ml-auto">
@@ -55,7 +55,7 @@
                         <a class="nav-link" href="mi-cuenta.html"><i class="fas fa-user-circle"></i> Mi Cuenta</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="login.html"><i class="fas fa-sign-out-alt"></i> Cerrar Sesion</a>
+                        <a class="nav-link" href="login.html"><i class="fas fa-sign-out-alt"></i> Cerrar Sesi&oacuten</a>
                     </li>
                 </ul>
             </div>
@@ -84,10 +84,10 @@
                 <thead class="thead-dark">
                     <tr>
                         <th scope="col">#</th>
-                        <th scope="col">Correo Electronico</th>
+                        <th scope="col">Correo Electr&oacutenico</th>
                         <th scope="col">Rol</th>
                         <th scope="col">Estado</th>
-                        <th scope="col" style="text-align: center;">Accion</th>
+                        <th scope="col" style="text-align: center;">Acci&oacuten</th>
                     </tr>
                 </thead>
                 <tbody>
@@ -138,7 +138,7 @@
                     <form autocomplete="off">
                         <div class="modal-body">
                             <div class="form-group col pl-0">
-                                <label for="inputCorreo">Correo Electronico</label>
+                                <label for="inputCorreo">Correo Electr&oacutenico</label>
                                 <input type="email" class="form-control" id="inputCorreo">
                             </div>
                             <div class="form-row">
@@ -177,7 +177,7 @@
                     <form autocomplete="off">
                         <div class="modal-body">
                             <div class="form-group col pl-0">
-                                <label for="inputCorreo">Correo Electronico</label>
+                                <label for="inputCorreo">Correo Electr&oacutenico</label>
                                 <input type="email" class="form-control" id="inputCorreo">
                             </div>
                             <div class="form-row">
@@ -200,7 +200,7 @@
                             </div>
                         </div>
                         <div class="modal-footer">
-                            <button type="reset" class="btn btn-secondary" data-dismiss="modal">Close</button>
+                            <button type="reset" class="btn btn-secondary" data-dismiss="modal">Cancelar</button>
                             <button type="submit" class="btn btn-primary">Agregar</button>
                         </div>
                     </form>
@@ -218,7 +218,7 @@
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.min.js" integrity="sha384-OgVRvuATP1z7JjHLkuOU7Xw704+h835Lr+6QL9UvYjZE3Ipu6Tp75j7Bh/kR0JKI" crossorigin="anonymous"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-datepicker/1.9.0/js/bootstrap-datepicker.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-datepicker/1.9.0/locales/bootstrap-datepicker.es.min.js"></script>
-    <script src="../js/formacion.js"></script>
+    <script src="/js/formacion.js"></script>
 </body>
 
 </html>
